# code-with-quarkus/src/main/java/org/acme/blog/OpenAPI.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Blog API",
    "description": "API für das Blog-System einschließlich Autoren, Blogs, Kommentare und Tags."
  },
  "paths": {
    "/author": {
      "get": {
        "operationId": "getAuthors",
        "description": "Erhalte alle Autoren.",
        "responses": {
          "200": {
            "description": "Liste aller Autoren.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Author"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "addAuthor",
        "description": "Erstellt einen neuen Autor. Authentifizierung erforderlich.",
        "requestBody": {
          "description": "Daten des neuen Autors.",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Author"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Autor erfolgreich erstellt.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Author"
                }
              }
            }
          },
          "400": {
            "description": "Ungültige Anfrage.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/author/{id}": {
      "put": {
        "operationId": "editAuthor",
        "description": "Bearbeitet einen bestehenden Autor.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "description": "Aktualisierte Daten des Autors.",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Author"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Autor erfolgreich bearbeitet.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Author"
                }
              }
            }
          },
          "404": {
            "description": "Autor nicht gefunden.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "400": {
            "description": "Ungültige Anfrage.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/blog": {
      "get": {
        "operationId": "getBlogs",
        "description": "Gibt alle Blogs zurück. Kann nach Autor, Titel oder Tag gefiltert werden.",
        "parameters": [
          {
            "name": "authorId",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "description": "ID des Autors zum Filtern der Blogs."
          },
          {
            "name": "title",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Titel zum Filtern der Blogs."
          },
          {
            "name": "tag",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Tag zum Filtern der Blogs."
          }
        ],
        "responses": {
          "200": {
            "description": "Liste aller Blogs.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Blog"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Keine Blogs gefunden.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "addBlog",
        "description": "Fügt einen neuen Blog hinzu.",
        "requestBody": {
          "description": "Daten des neuen Blogs.",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Blog"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Blog erfolgreich erstellt.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Blog"
                }
              }
            }
          },
          "400": {
            "description": "Ungültige Anfrage.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/blog/{id}": {
      "delete": {
        "operationId": "deleteBlog",
        "description": "Löscht einen Blog anhand der ID. Authentifizierung und Admin-Rolle erforderlich.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Blog erfolgreich gelöscht."
          },
          "404": {
            "description": "Blog nicht gefunden.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/blog/{blogId}/comments": {
      "get": {
        "operationId": "getComments",
        "description": "Gibt alle Kommentare für einen bestimmten Blog zurück.",
        "parameters": [
          {
            "name": "blogId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID des Blogs, um Kommentare abzurufen."
          }
        ],
        "responses": {
          "200": {
            "description": "Liste aller Kommentare für den Blog.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Comment"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Keine Kommentare gefunden.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "addComment",
        "description": "Fügt einen neuen Kommentar zu einem Blog hinzu.",
        "parameters": [
          {
            "name": "blogId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID des Blogs, zu dem der Kommentar hinzugefügt wird."
          }
        ],
        "requestBody": {
          "description": "Daten des neuen Kommentars.",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Comment"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Kommentar erfolgreich erstellt.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Comment"
                }
              }
            }
          },
          "404": {
            "description": "Blog nicht gefunden.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "400": {
            "description": "Ungültige Anfrage.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/comments/{id}": {
      "delete": {
        "operationId": "deleteComment",
        "description": "Löscht einen Kommentar anhand der ID.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Kommentar erfolgreich gelöscht."
          },
          "404": {
            "description": "Kommentar nicht gefunden.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/tag": {
      "get": {
        "operationId": "getTags",
        "description": "Gibt alle Tags zurück.",
        "responses": {
          "200": {
            "description": "Liste aller Tags.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Tag"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "addTag",
        "description": "Fügt einen neuen Tag hinzu.",
        "requestBody": {
          "description": "Daten des neuen Tags.",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Tag"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tag erfolgreich erstellt.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Tag"
                }
              }
            }
          },
          "400": {
            "description": "Ungültige Anfrage.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/blog/{blogId}/tags": {
      "post": {
        "operationId": "addTagToBlog",
        "description": "Fügt einen Tag zu einem Blog hinzu.",
        "parameters": [
          {
            "name": "blogId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID des Blogs, zu dem der Tag hinzugefügt wird."
          }
        ],
        "requestBody": {
          "description": "Daten des Tags.",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Tag"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tag erfolgreich zum Blog hinzugefügt.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Tag"
                }
              }
            }
          },
          "404": {
            "description": "Blog nicht gefunden.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "400": {
            "description": "Ungültige Anfrage.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Author": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "biography": {
            "type": "string"
          }
        },
        "required": ["name"]
      },
      "Blog": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "author": {
            "$ref": "#/components/schemas/Author"
          },
          "tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Tag"
            }
          }
        },
        "required": ["title", "content", "author"]
      },
      "Comment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "content": {
            "type": "string"
          },
          "blog": {
            "$ref": "#/components/schemas/Blog"
          }
        },
        "required": ["content", "blog"]
      },
      "Tag": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          }
        },
        "required": ["name"]
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        }
      }
    }
  }
}